+++ contrib/packs/actions/pack_mgmt/virtualenv_setup_prerun.py
@@ -6,10 +6,11 @@ from st2actions.runners.pythonrunner import Action
 class PacksTransformationAction(Action):
     def run(self, packs_status):
         """
-        :param packs: A list of packs to create the environment for.
-        :type: packs: ``list``
+        :param packs_status: Result from packs.download action.
+        :type: packs_status: ``dict``
         """
         packs = []
         for pack_name, status in six.iteritems(packs_status):
-            packs.append(pack_name)
+            if 'success' in status.lower():
+                packs.append(pack_name)
         return packs