+++ st2auth/st2auth/config.py
@@ -17,6 +17,10 @@ from oslo.config import cfg
 
 from st2common import config as st2cfg
 from st2common.constants.system import VERSION_STRING
+from st2common.constants.auth import DEFAULT_MODE
+from st2common.constants.auth import DEFAULT_BACKEND
+from st2common.constants.auth import VALID_MODES
+from st2auth.backends import VALID_BACKEND_NAMES
 
 
 def parse_args(args=None):
@@ -29,19 +33,33 @@ def _register_common_opts():
 
 def _register_app_opts():
     auth_opts = [
-        cfg.StrOpt('host', default='0.0.0.0'),
-        cfg.IntOpt('port', default=9100),
-        cfg.StrOpt('cert', default='/etc/apache2/ssl/mycert.crt'),
-        cfg.StrOpt('key', default='/etc/apache2/ssl/mycert.key'),
-        cfg.StrOpt('logging', default='conf/logging.conf'),
-        cfg.BoolOpt('debug', default=False)]
-    cfg.CONF.register_opts(auth_opts, group='auth')
+        cfg.StrOpt('host', default='0.0.0.0', help='Host on which the service should listen on.'),
+        cfg.IntOpt('port', default=9100, help='Port on which the service should listen on.'),
+        cfg.BoolOpt('use_ssl', default=False, help='Specify to enable SSL / TLS mode'),
+        cfg.StrOpt('cert', default='/etc/apache2/ssl/mycert.crt',
+                   help='Path to the SSL certificate file. Only used when "use_ssl" is specified.'),
+        cfg.StrOpt('key', default='/etc/apache2/ssl/mycert.key',
+                   help='Path to the SSL private key file. Only used when "use_ssl" is specified.'),
+        cfg.StrOpt('logging', default='conf/logging.conf',
+                   help='Path to the logging config.'),
+        cfg.BoolOpt('debug', default=False,
+                    help='Specify to enable debug mode.'),
+        cfg.StrOpt('mode', default=DEFAULT_MODE,
+                   help='Authentication mode (%s)' % (','.join(VALID_MODES))),
+        cfg.StrOpt('backend', default=DEFAULT_BACKEND,
+                   help='Authentication backend to use in a standalone mode (%s).' %
+                   (','.join(VALID_BACKEND_NAMES))),
+        cfg.StrOpt('backend_kwargs', default=None,
+                   help='JSON serialized arguments which are passed to the authentication backend'
+                        ' in a standalone mode.'),
+    ]
+    cfg.CONF.register_cli_opts(auth_opts, group='auth')
 
     api_opts = [
         cfg.ListOpt('allow_origin', default=['http://localhost:3000'],
             help='List of origins allowed'),
     ]
-    cfg.CONF.register_opts(api_opts, group='api')
+    cfg.CONF.register_cli_opts(api_opts, group='api')
 
 
 def register_opts():