+++ st2reactor/st2reactor/sensor/base.py
@@ -73,18 +73,12 @@ class PollingSensor(BaseSensor):
     """
     Base class to be inherited from by the active sensors.
 
-    Active sensors are the ones which periodically polling a 3rd party system
-    for new information.
+    Active sensors periodically poll a 3rd party system for new information.
     """
 
     def __init__(self, sensor_service, config, poll_interval=5):
         super(PollingSensor, self).__init__(sensor_service=sensor_service, config=config)
         self._poll_interval = poll_interval
-
-    def run(self):
-        while True:
-            self.poll()
-            eventlet.sleep(self._poll_interval)
 
     @abc.abstractmethod
     def poll(self):
@@ -92,3 +86,26 @@ class PollingSensor(BaseSensor):
         Poll 3rd party system for new information.
         """
         pass
+
+    def run(self):
+        while True:
+            self.poll()
+            eventlet.sleep(self._poll_interval)
+
+    def get_poll_interval(self):
+        """
+        Retrieve current poll interval.
+
+        :return: Current poll interval.
+        :rtype: ``float``
+        """
+        return self._poll_interval
+
+    def set_poll_interval(self, poll_interval):
+        """
+        Set the poll interval.
+
+        :param poll_interval: Poll interval to use.
+        :type poll_interval: ``float``
+        """
+        self._poll_interval = poll_interval